:recycle: refactoring - update file & directory name / remove unuse img files
