+++ src/components/ContentModal.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { useSelector , useDispatch } from 'react-redux';
 import styled from 'styled-components';
 import ContentSlider from './ContentSlider';
-import {likeButton, kakaoLikeButton} from '../modules/likeReducer'
+import {likeButton, kakaoLikeButton} from '../modules/like'
 import Loading from './Loading';
 import { __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED } from 'react-dom';
